requirements.txt
line-bot-sdk
flask
requests
datetime
beautifulsoup4
pillow
psycopg[binary]